fix(runtime): validate pipeline webhook identity

+++ runtime/internal/trigger/bundle.go
@@ -28,13 +28,13 @@ func mustCompileBundleSchema() *jsonschema.Schema {
 	return c.MustCompile("bundle.schema.json")
 }
 
-// Bundle 是 bundle-g{sha}.json 的解析结果(contracts/bundle.schema.json)。
+// Bundle 是 bundle-g{sha}-p{global_id}.json 的解析结果(contracts/bundle.schema.json)。
 type Bundle struct {
 	BundleVersion    int             `json:"bundle_version"`
 	Project          string          `json:"project"`
 	Commit           string          `json:"commit"`
 	PipelineID       int             `json:"pipeline_id"`
-	PipelineGlobalID int             `json:"pipeline_global_id"`
+	PipelineGlobalID int64           `json:"pipeline_global_id"`
 	Version          string          `json:"version"`
 	CreatedAt        string          `json:"created_at"`
 	Packages         []wf.PackageRef `json:"packages"`
